+++ apps/discordbot/src/events/index.ts
@@ -1,4 +1,4 @@
 export * from './ready.event';
 export * from './warn.event';
-export * from './message.event';
+export * from './message-create.event';
 export * from './guild-member-add.event';
